exercises(sum-of-multiples): use u32, not usize

+++ exercises/practice/sum-of-multiples/.meta/example.zig
@@ -1,11 +1,11 @@
 const std = @import("std");
 const mem = std.mem;
 
-pub fn sum(allocator: mem.Allocator, factors: []const usize, limit: usize) !usize {
-    var intset = std.AutoHashMap(usize, void).init(allocator);
+pub fn sum(allocator: mem.Allocator, factors: []const u32, limit: u32) !u32 {
+    var intset = std.AutoHashMap(u32, void).init(allocator);
     defer intset.deinit();
 
-    var result: usize = 0;
+    var result: u32 = 0;
     for (factors) |f| {
         if (f != 0) {
             var n = f;
